fix: change client-component to server-component
