+++ package.json
@@ -45,7 +45,7 @@
         "@types/react": "^18",
         "@types/react-dom": "^18",
         "@vanilla-extract/next-plugin": "^2.4.1",
-        "eslint": "^9.13.0",
+        "eslint": "^8",
         "eslint-config-next": "14.2.3",
         "eslint-config-prettier": "^9.1.0",
         "eslint-plugin-import": "^2.29.1",
