+++ src/features/home/components/card-container/styles.css.ts
@@ -1,0 +1,10 @@
+import { style } from "@vanilla-extract/css"
+
+const styles = {
+    container: style({
+        width: 202,
+        height: "auto"
+    })
+}
+
+export default styles
