--- src/features/home/components/cardContainer/styles.css.ts
@@ -1,11 +1,0 @@
-import { style } from "@vanilla-extract/css"
-
-const styles = {
-    link: style({
-        textDecoration: "none",
-        height: '100%',
-        display: 'block'
-    })
-}
-
-export default styles
