+++ src/features/home/components/icon/icon.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import Image from "next/image"
 import styles from "./style.css"
 import ICON from "@/assets/home/icon.jpg"
